+++ CHANGELOG.md
@@ -31,3 +31,4 @@
 - Implement `xsd:decimal` codec
 - Implement `xsd:double` codec
 - Implement `xsd:float` codec
+- Implement `xsd:unsignedInt` codec